move everything to Book and add save function to be able to do more things in memory / not on synced drive
this still doesn't actually solve the issue of failing when expandrive is syncing. I guess it's a problem with having an async fs...at some point you're downloading changes, so it could be partial. it doesn't seem like this has to be the case very often, though, and expandrive is just being weird in this case. regardless, it's an issue. maybe i can find a different drive sync program?